Refactor: add a new file for editor navigation extension
