+++ WordPress/Classes/System/RootViewPresenter+EditorNavigation.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+extension RootViewPresenter {
+    func currentOrLastBlog() -> Blog? {
+        if let blog = currentlyVisibleBlog() {
+            return blog
+        }
+        let context = ContextManager.shared.mainContext
+        return Blog.lastUsedOrFirst(in: context)
+    }
+}
