+++ WordPress/Classes/System/RootViewPresenter.swift
@@ -43,13 +43,3 @@ protocol RootViewPresenter {
     func popNotificationsTabToRoot()
 
 }
-
-extension RootViewPresenter {
-    func currentOrLastBlog() -> Blog? {
-        if let blog = currentlyVisibleBlog() {
-            return blog
-        }
-        let context = ContextManager.shared.mainContext
-        return Blog.lastUsedOrFirst(in: context)
-    }
-}
